Natalis-ML-API-main/requirements.txt
flask
numpy
pillow
torch
torchvision
openpyxl
pandas
scipy
opencv-python-headless
matplotlib
segmentation-models-pytorch
scikit-image